Game Manager and Raycast
Add Game Manager to handle global variables from Score Manager and Level Manager using Singleton pattern.
Add Score Manager to handle player scoring when killing an enemy and collecting items.
Modify EnemyController so that it has a switch that controls the amount of points given based on the enemy type.
Add Level Manager to handle change of scenes on collision with LevelEnd trigger.
Implement Spherecast on EnemyController so that enemies only move and throw rocks when the player is in sight.
Implement Raycast on PlayerController so that the player only jumps when touching the ground.